requirements.txt
streamlit
flask
google-generativeai
plotly
pandas
requests
# gunicorn